refactor lambda function
