+++ agrialexa/alexa.go
@@ -1,0 +1,75 @@
+package agrialexa
+
+import (
+	"context"
+	"errors"
+	"fmt"
+	"log"
+
+	"github.com/ericdaugherty/alexa-skills-kit-golang"
+	"github.com/hatobus/UKEMOCHI/outbound"
+)
+
+type SmartAgri struct{}
+
+func (s *SmartAgri) OnSessionStarted(ctx context.Context, request *alexa.Request, session *alexa.Session, ctxPtr *alexa.Context, response *alexa.Response) error {
+
+	log.Printf("OnSessionStarted requestId=%s, sessionId=%s", request.RequestID, session.SessionID)
+
+	return nil
+}
+
+// OnLaunch called with a reqeust is received of type LaunchRequest
+func (s *SmartAgri) OnLaunch(ctx context.Context, request *alexa.Request, session *alexa.Session, ctxPtr *alexa.Context, response *alexa.Response) error {
+	speechText := "これはスマートアグリの情報を取得できます。"
+	speechText = speechText + "取得できる情報は 温度、湿度、二酸化炭素濃度、水分量に照度です。"
+	speechText = speechText + "一から三号機までの情報にそれぞれアクセスできます。"
+
+	log.Printf("OnLaunch requestId=%s, sessionId=%s", request.RequestID, session.SessionID)
+
+	// response.SetSimpleCard(cardTitle, speechText)
+	response.SetOutputText(speechText)
+	response.SetRepromptText(speechText)
+
+	response.ShouldSessionEnd = true
+
+	return nil
+}
+
+// OnIntent called with a reqeust is received of type IntentRequest
+func (s *SmartAgri) OnIntent(ctx context.Context, request *alexa.Request, session *alexa.Session, ctxPtr *alexa.Context, response *alexa.Response) error {
+
+	log.Printf("OnIntent requestId=%s, sessionId=%s, intent=%s", request.RequestID, session.SessionID, request.Intent.Name)
+
+	switch request.Intent.Name {
+	case "getParamIntent":
+		log.Println("getParamIntent triggered")
+		speechText, err := outbound.Getsmartagriinfo(request.Intent.Slots)
+		if err != nil {
+			fmt.Println(err)
+			speechText = "すみません、情報を取得できませんでした。"
+		}
+
+		response.SetOutputText(speechText)
+
+		log.Printf("Set Output speech, value now: %s", response.OutputSpeech.Text)
+	case "AMAZON.HelpIntent":
+		log.Println("何か助けが必要ですか")
+		speechText := "何か助けが必要ですか"
+
+		response.SetOutputText(speechText)
+		response.SetRepromptText(speechText)
+	default:
+		return errors.New("Invalid Intent")
+	}
+
+	return nil
+}
+
+// OnSessionEnded called with a reqeust is received of type SessionEndedRequest
+func (s *SmartAgri) OnSessionEnded(ctx context.Context, request *alexa.Request, session *alexa.Session, ctxPtr *alexa.Context, response *alexa.Response) error {
+
+	log.Printf("OnSessionEnded requestId=%s, sessionId=%s", request.RequestID, session.SessionID)
+
+	return nil
+}
